requirements.txt
environ==1.0
discord.py==2.0.1
